worker: (broken) move action and worker to internal/ for cleaner git history
