+++ cmd/tink-worker/internal/action.go
@@ -1,4 +1,4 @@
-package main
+package internal
 
 import (
 	"bufio"
