+++ cmd/tink-worker/internal/worker.go
@@ -1,4 +1,4 @@
-package main
+package internal
 
 import (
 	"context"
